chore(swaps): separate BoltzSwapAction enum. Make BoltzOperationClassifier.cs static class
