+++ NArk.Swaps/Boltz/BoltzSwapAction.cs
@@ -1,0 +1,11 @@
+namespace NArk.Swaps.Boltz;
+
+public enum BoltzSwapAction
+{
+    CanCoopRefundSubmarine,
+    CanCoopRefundArkToBtc,
+    CanCoopRefundBtcToArk,
+    CanRenegotiateChain,
+    CanClaimChain,
+    ReadyToSignClaim,
+}
